External Sort using Seastar - Pass 1
This commit includes a working pass 1 of an external merge sort using
the seastar framework. In this first pass, the input file is split
evenly across the available shards and then they are further subdivided
into batches based on the available memory in the shards. The records in
these batches are then sorted within themselves and written to disk.

In a followup, these separate batches has to be merged to form a single
sorted file.