External Sort using Seastar - Pass 2
This commit adds a pass 2 in which each shard merges all the sorted
batches produced by it in the first pass into a single sorted file.

Once the files are merged per shard, the same second pass service is
called again on a single shard to merge all the files from across the
shards into a single sorted output file.